Add HoustList Controller
