+++ UserAuth/Controllers/UserAuthController.cs
@@ -65,7 +65,6 @@ namespace UserAuth.Controllers
                             InsertNewAccount.job = user.job;
                             InsertNewAccount.gender = user.gender;
                             InsertNewAccount.role = user.role;
-                            InsertNewAccount.userIntro = user.userIntro;
                             InsertNewAccount.userIntro = userIntro;
 
                             db.UserEntities.Add(InsertNewAccount);
